made read components reusable
