+++ apps/ethereum/token/components/WriteFunctions/CreateInstance.tsx
@@ -1,11 +1,10 @@
 
 import { useState } from 'react';
-import { styled } from '../../stitches.config';
 import * as Accordion from '@radix-ui/react-accordion';
 import { useEthereum } from '@decentology/hyperverse-ethereum';
 import { useToken } from '../../source/useToken';
-
-
+import { Box, Item, TriggerContainer, Trigger, Parameters, Input, Content, Button} from './WriteComponents'
+ 
 const CreateInstance = () => {
   const { address } = useEthereum();
   const { NewInstance } = useToken();
@@ -35,7 +34,7 @@ const CreateInstance = () => {
       <h4>New Instance</h4>
       <p>Create your own instance of a token </p>
       <Accordion.Root type="single" collapsible>
-      <Instance value='item-1'>
+      <Item value='item-1'>
         <TriggerContainer>
           <Trigger>
               New Instance
@@ -49,7 +48,7 @@ const CreateInstance = () => {
           <Button disabled={!address} onClick={createNewInstance}>{!address ? 'Connet Wallet' : 'Create Instance'}</Button>
         </Content>
         </Parameters>
-      </Instance>
+      </Item>
     </Accordion.Root>
     </Box>
     
@@ -57,86 +56,3 @@ const CreateInstance = () => {
 }
 
 export default CreateInstance;
-
-const Box = styled('div', {
-  minHeight: '150px',
-  maxWidth: '220px',
-  display: 'flex',
-  flexDirection: 'column',
-  backgroundColor: '$blue500',
-  borderRadius:'10px',
-  color: 'white',
-  padding: '30px 20px',
-  alignItems: 'center',
-  '& h4': {
-    marginTop: '10px',
-    fontSize: '1.2rem',
-    fontWeight: '500',
-  },
-  '& p' : {
-    margin: '10px 0 30px',
-    fontSize: '0.8rem',
-  }
-})
-
-
-const Button = styled('button', {
-  maxWidth: '150px',
-	backgroundColor: '$yellow100',
-	outline: 'none',
-	border: 'none',
-	padding: '10px 15px',
-	borderRadius: '90px',
-	cursor: 'pointer',
-  margin: '10px auto -5px',
-  '&:hover': {
-    opacity: 0.8,
-  }
-})
-
-const TriggerContainer = styled(Accordion.Header, {
-  margin: '0 !important',
-})
-
-
-const Trigger = styled(Accordion.Trigger, {
-  minWidth: '130px',
-	backgroundColor: '$blue200',
-	outline: 'none',
-	border: 'none',
-	padding: '10px 15px',
-	borderRadius: '90px',
-	cursor: 'pointer',
-  margin: '10px auto 0',
- 
-})
-
-const Parameters = styled(Accordion.Content, {
-  display: 'flex',
-  flexDirection: 'column',
-})
-
-const Content = styled('div', {
-  display: 'flex',
-  flexDirection: 'column',
-  width: '100%',
-  marginTop: '10px',
-})
-
-const Input = styled('input', {
-  width: '200px',
-  margin: '3px 0',
-  padding: '5px',
-  backgroundColor: '$gray200',
-  border: 'none',
-  color: 'white',
-
-})
-
-const Instance = styled(Accordion.Item, {
-  display: 'flex',
-  flexDirection: 'column',
-  justifyContent: 'center',
-  alignItems: 'center',
-  
-})
